+++ src/sections/finance/components/finance-deposit-from-metamask.tsx
@@ -1,20 +1,15 @@
-// React and libraries imports
-import { FC, useEffect, useState } from 'react';
+// REACT IMPORTS
+import { FC } from 'react';
+
+// REDUX IMPORTS
 import { useSelector } from 'react-redux';
-import { Address } from 'viem';
 
-// @MUI Imports
-import Button from '@mui/material/Button';
-
-// Project Imports
-import Iconify from '@src/components/iconify';
-import { connectToMetaMask } from '@src/utils/metamask';
+// LOCAL IMPORTS
+import { useMetaMask } from '@src/hooks/use-metamask';
 import { useDepositMetamask } from '@src/hooks/use-deposit-metamask';
-import { LoadingScreen } from '@src/components/loading-screen';
-import FinanceDeposit from './finance-deposit';
-import { ERRORS } from '@notifications/errors.ts';
-import { notifyError } from '@notifications/internal-notifications.ts';
-import { Box } from '@mui/system';
+import FinanceDeposit from '@src/sections/finance/components/finance-deposit';
+import FinanceMetamaskLoader from '@src/sections/finance/components/finance-metamask-loader.tsx';
+import FinanceMetamaskButton from '@src/sections/finance/components/finance-metamask-button.tsx';
 
 interface FinanceDepositFromMetamaskProps {
   onClose: () => void;
@@ -22,65 +17,13 @@ interface FinanceDepositFromMetamaskProps {
 
 const FinanceDepositFromMetamask: FC<FinanceDepositFromMetamaskProps> = ({ onClose }) => {
   const sessionData = useSelector((state: any) => state.auth.session);
-
-  const [address, setAddress] = useState<Address | undefined>();
-  const [connecting, setConnecting] = useState(false);
-
-  // Specific hook for Metamask
   const depositHook = useDepositMetamask();
+  const { address, connecting, connect } = useMetaMask();
 
-  // Try reconnecting if the user connected MetaMask before
-  useEffect(() => {
-    const walletConnected = localStorage.getItem('walletConnected');
-    if (walletConnected === 'true') {
-      handleConnectMetaMask();
-    }
-  }, []);
+  if (connecting) return <FinanceMetamaskLoader />;
+  if (!address) return <FinanceMetamaskButton connect={connect} />;
 
-  const handleConnectMetaMask = async () => {
-    setConnecting(true);
-    try {
-      const walletAddress = await connectToMetaMask();
-      setAddress(walletAddress);
-      localStorage.setItem('walletConnected', 'true');
-    } catch (err) {
-      notifyError(ERRORS.METAMASK_CONNECTING_ERROR);
-    } finally {
-      setConnecting(false);
-    }
-  };
-
-  if (connecting) {
-    return (
-      <Box sx={{ mx: 4, my: 8 }}>
-        <LoadingScreen />
-      </Box>
-    );
-  }
-
-  // If the wallet is NOT connected, show a button
-  if (!address) {
-    return (
-      <Button
-        sx={{ m: 4, p: 1.5 }}
-        startIcon={<Iconify icon="logos:metamask-icon" />}
-        variant="outlined"
-        onClick={handleConnectMetaMask}
-      >
-        Connect MetaMask
-      </Button>
-    );
-  }
-
-  // If the wallet IS connected, render the deposit component
-  return (
-    <FinanceDeposit
-      address={address}
-      recipient={sessionData?.address}
-      depositHook={depositHook}
-      onClose={onClose}
-    />
-  );
+  return <FinanceDeposit address={address} recipient={sessionData?.address} depositHook={depositHook} onClose={onClose} />;
 };
 
 export default FinanceDepositFromMetamask;
