+++ src/sections/finance/components/finance-withdraw-from-metamask.tsx
@@ -1,83 +1,27 @@
-// React and libraries imports
-import { FC, useEffect, useState } from 'react';
-import { Address } from 'viem';
+// REACT IMPORTS
+import { FC } from 'react';
 
-// @MUI Imports
-import Button from '@mui/material/Button';
-
-// Import Components
-import Iconify from '@src/components/iconify';
-import FinanceWithdraw from './finance-withdraw.tsx';
+// LOCAL IMPORTS
+import { useMetaMask } from '@src/hooks/use-metamask';
 import { useWithdrawMetamask } from '@src/hooks/use-withdraw-metamask';
 import { useGetVaultBalance } from '@src/hooks/use-get-vault-balance';
-import { connectToMetaMask } from '@src/utils/metamask';
-import { LoadingScreen } from '@src/components/loading-screen';
-
-// Notifications
-import { notifyError } from '@src/utils/notifications/internal-notifications';
-import { ERRORS } from '@src/utils/notifications/errors';
-import { Box } from '@mui/system';
+import FinanceWithdraw from '@src/sections/finance/components/finance-withdraw';
+import FinanceMetamaskLoader from '@src/sections/finance/components/finance-metamask-loader.tsx';
+import FinanceMetamaskButton from '@src/sections/finance/components/finance-metamask-button.tsx';
 
 interface FinanceWithdrawFromMetamaskProps {
   onClose: () => void;
 }
 
 const FinanceWithdrawFromMetamask: FC<FinanceWithdrawFromMetamaskProps> = ({ onClose }) => {
-  const [address, setAddress] = useState<Address | undefined>();
-  const [connecting, setConnecting] = useState(false);
-  const { balance } = useGetVaultBalance(address);
   const withdrawHook = useWithdrawMetamask();
+  const { address, connecting, connect } = useMetaMask();
+  const { balance } = useGetVaultBalance(address);
 
-  // Auto-reconnect if MetaMask was previously connected
-  useEffect(() => {
-    const walletConnected = localStorage.getItem('walletConnected');
-    if (walletConnected === 'true') {
-      handleConnectMetaMask();
-    }
-  }, []);
+  if (connecting) return <FinanceMetamaskLoader />;
+  if (!address) return <FinanceMetamaskButton connect={connect} />;
 
-  const handleConnectMetaMask = async () => {
-    setConnecting(true);
-    try {
-      const walletAddress = await connectToMetaMask();
-      setAddress(walletAddress);
-      localStorage.setItem('walletConnected', 'true');
-    } catch (err) {
-      notifyError(ERRORS.METAMASK_CONNECTING_ERROR);
-    } finally {
-      setConnecting(false);
-    }
-  };
-
-  if (connecting) {
-    return (
-      <Box sx={{ mx: 4, my: 8 }}>
-        <LoadingScreen />
-      </Box>
-    );
-  }
-
-  if (!address) {
-    return (
-      <Button
-        sx={{ m: 4, p: 1.5 }}
-        startIcon={<Iconify icon="logos:metamask-icon" />}
-        variant="outlined"
-        onClick={handleConnectMetaMask}
-      >
-        Connect MetaMask
-      </Button>
-    );
-  }
-
-  return (
-    <FinanceWithdraw
-      address={address}
-      withdrawHook={withdrawHook}
-      balance={balance}
-      onClose={onClose}
-    />
-  );
+  return <FinanceWithdraw address={address} withdrawHook={withdrawHook} balance={balance} onClose={onClose} />;
 };
 
 export default FinanceWithdrawFromMetamask;
