+++ src/sections/finance/components/finance-withdraw-modal.tsx
@@ -61,7 +61,7 @@ export const FinanceWithdrawModal: FC<FinanceWithdrawModalProps> = ({ open, onCl
           zIndex: 9,
           borderBottom: '1px solid rgba(255, 255, 255, 0.08)',
           [`& .${tabsClasses.flexContainer}`]: { justifyContent: 'center', px: 1 },
-          [`& .${tabsClasses.scroller}`]: { display: 'flex' },
+          [`& .${tabsClasses.scroller}`]: { display: 'flex', justifyContent: 'center' },
         }}
       >
         {TABS.map((tab) => (
